Updating Hex tiles to allow for walls on edges of rooms / connecting hallways

+++ src/level-generation/level_generator.h
@@ -95,23 +95,23 @@ public:
   int GetNumRooms();
 
 private:
-	void m_GenerateRoom(Vector<uint8_t> &tile_map, HashMap<String, Ref<Tile>> *grid_of_tiles, TileGrid *root, Vector<Ref<Tile>> &spawnable_locations);
+	void m_GenerateRoom(Vector<uint16_t> &tile_map, HashMap<String, Ref<Tile>> *grid_of_tiles, TileGrid *root, Vector<Ref<Tile>> &spawnable_locations);
   m_Rooms_Graph* m_GenerateRoomGraph(Vector2i starting_location, String file);
   void m_ReplaceNodesInPattern(m_Rooms_Graph *rooms_graph);
-  void m_GenerateGraphTileBitMap(Vector<uint8_t> &tile_bit_map, m_Rooms_Graph *graph, Vector2i grid_origin);
-  void m_ConnectGraphNodes(Vector<uint8_t> &tile_bilt_map, m_Rooms_Graph *graph);
-  Ref<Tile> m_InstantiateTile(Object* GenerationComunicator, Vector<Ref<Tile>> &spawnable_locations, Vector3 location, int q, int r, int tile_type);
-  void m_SetTileMeshAndMaterial(TileMeshGenerator *mesh_generator, ResourceLoader *rl, String mesh_material_name, String tile_mesh_name, int tile_type);
+  void m_GenerateGraphTileBitMap(Vector<uint16_t> &tile_bit_map, m_Rooms_Graph *graph, Vector2i grid_origin);
+  void m_ConnectGraphNodes(Vector<uint16_t> &tile_bilt_map, m_Rooms_Graph *graph);
+  Ref<Tile> m_InstantiateTile(Object* GenerationComunicator, Vector<Ref<Tile>> &spawnable_locations, Vector3 location, int q, int r, uint16_t tile_type);
+  void m_SetTileMeshAndMaterial(TileMeshGenerator *mesh_generator, ResourceLoader *rl, String mesh_material_name, String tile_mesh_name, uint16_t tile_type);
   Vector<int> m_GenerateInteractableType(int num_points);
 	Vector<Vector2i> m_GenerateMST(const Vector<Vector2i> &room_centers, m_Room_Tree_Node *root, u_int8_t size);
 	void m_GenerateNeighborsForNode(m_Room_Tree_Node *current_node, m_Room_Tree_Node *root, Vector<Vector2i> &neighbor_list, int level);
 	m_Best_Neighbors m_FindNearest(m_Room_Tree_Node *node, Vector2i goal_room, m_Best_Neighbors best_neighbor, int level);
-	m_Room_Tree_Node *m_GenerateTileBitMap(Vector<uint8_t> &tile_bit_map, m_Room_Tree_Node *root_room /*Vector<Vector2i> &room_centers*/, int &num_of_rooms_remaining,
+	m_Room_Tree_Node *m_GenerateTileBitMap(Vector<uint16_t> &tile_bit_map, m_Room_Tree_Node *root_room /*Vector<Vector2i> &room_centers*/, int &num_of_rooms_remaining,
 			int current_level, int max_level, Vector2i max_grid_size);
-	bool m_OverlappingRooms(const Vector<uint8_t> &tile_bit_map, Vector2i center, int radius);
-	void m_FillBitMap(Vector<uint8_t> &tile_bit_map, int q_center, int r_center, int radius);
-	void m_ConnectTiles(Vector<uint8_t> &tile_bit_map, Vector<Vector2i> room_neighbors);
-	void m_DrawLineTiles(Vector<uint8_t> &tile_bit_map, Vector2i first_room_center, Vector2i second_room_center);
+	bool m_OverlappingRooms(const Vector<uint16_t> &tile_bit_map, Vector2i center, int radius);
+	void m_FillBitMap(Vector<uint16_t> &tile_bit_map, int q_center, int r_center, int radius);
+	void m_ConnectTiles(Vector<uint16_t> &tile_bit_map, Vector<Vector2i> room_neighbors);
+	void m_DrawLineTiles(Vector<uint16_t> &tile_bit_map, Vector2i first_room_center, Vector2i second_room_center);
 	void m_AddNodeToTree(m_Room_Tree_Node *root_room, Vector2i new_room, int level);
 	Vector2i m_HexRound(Vector2i first_room, Vector2i second_room, int distance, int step);
 	int m_HexDistance(Vector2i first_room, Vector2i second_room);
