+++ src/tiles/obstacle.h
@@ -7,7 +7,7 @@ class Obstacle : public Tile {
 	GDCLASS(Obstacle, Tile)
 public:
   Obstacle();
-	Obstacle(Vector3 position, int r, int c, bool flat_topped, float outer_size, float inner_size, float height, uint8_t tile_type);
+	Obstacle(Vector3 position, int r, int c, bool flat_topped, float outer_size, float inner_size, float height, uint16_t tile_type);
   ~Obstacle();
 
   protected:
