+++ src/tiles/starting_tile.cpp
@@ -8,7 +8,7 @@ godot::StartingTile::StartingTile() : Tile() {
 }
 
 
-godot::StartingTile::StartingTile(Vector3 position, int r, int c, bool flat_topped, float outer_size, float inner_size, float height, uint8_t type) : Tile(position, r, c, flat_topped, outer_size, inner_size, height, type) {
+godot::StartingTile::StartingTile(Vector3 position, int r, int c, bool flat_topped, float outer_size, float inner_size, float height, uint16_t type) : Tile(position, r, c, flat_topped, outer_size, inner_size, height, type) {
 
 }
 void godot::StartingTile::SpawnCharacter() {
