+++ src/tiles/starting_tile.h
@@ -9,7 +9,7 @@ class StartingTile : public Tile, public IStepOnTile {
 
 public:
 	StartingTile();
-	StartingTile(Vector3 position, int r, int c, bool flat_topped, float outer_size, float inner_size, float height, uint8_t type);
+	StartingTile(Vector3 position, int r, int c, bool flat_topped, float outer_size, float inner_size, float height, uint16_t type);
 	~StartingTile();
 
 	void TileSteppedOnEvent() override;
