+++ src/tiles/unit_spawner.h
@@ -8,7 +8,7 @@ class UnitSpawner : public Tile  {
 
 public:
   UnitSpawner();
-	UnitSpawner(Vector3 position, int r, int c, bool flat_topped, float outer_size, float inner_size, float height, uint8_t tile_type);
+	UnitSpawner(Vector3 position, int r, int c, bool flat_topped, float outer_size, float inner_size, float height, uint16_t tile_type);
   ~UnitSpawner();
 
   void SpawnCharacter();
